Simple setup software copy
Probably the best part of this is figuring our which drive let the script starts from.  I needed that because the script might have started from any old drive letter. And of course I avoided using it as much as possible.

The lines determining what kind of windows was running was marginally helpful, but I just copied that off StackOverflow.