Simple copy of a default software set
Just copies software from a USB or other media to a user's drive.

This is for minor stuff that did not get sent in with the image.

I like the ability to know what drive I was running from, but I tried not to reference file by drive letter if I could help it.

The ability to know whether I was on a 32 or 64 system would have been nice, but we went all 64 bit. I think I copied that line from StackOverflow.